pyproject.toml
[tool.poetry]
name = "naptwitter"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.112.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.32"}
uvicorn = "^0.30.5"
alembic = "^1.13.2"
python-dotenv = "^1.0.1"
asyncpg = "^0.29.0"
loguru = "^0.7.2"
psycopg2-binary = "^2.9.9"
fastapi-cache2 = {extras = ["redis"], version = "^0.2.2"}
pytest = "^8.3.2"
pytest-asyncio = "^0.24.0"
python-multipart = "^0.0.9"
black = "^24.8.0"
celery = "^5.4.0"
rabbitmq = "^0.2.0"
flower = "^2.0.1"
pydantic = {extras = ["email"], version = "^2.9.1"}
fastapi-users = {extras = ["sqlalchemy"], version = "^13.0.0"}
fastapi-mail = {extras = ["aioredis"], version = "^1.4.1"}
passlib = "^1.7.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode="auto"